Fixed bug where login sent user to home screen
