+++ app/src/main/java/com/thavelka/feedme/ListingsFragment.java
@@ -128,7 +128,7 @@ public class ListingsFragment extends Fragment {
             try {
                 mListings = query.find();
                 Log.d(TAG, "got " + mListings.size() + " objects");
-                mAdapter = new ParseAdapter(getActivity(), mListings, false);
+                mAdapter = new MainListingAdapter(getActivity(), mListings, false);
                 return mListings;
             } catch (ParseException e) {
                 e.printStackTrace();
